% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.gam.R
\name{confint.gam}
\alias{confint.gam}
\title{Confidence Intervals for Model Parameters}
\usage{
\method{confint}{gam}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object of class \code{"gam"}.}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing, all
parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{not implemented}
}
\value{
A tidy data frame containing parameter names, estimates and
confidence intervals for parametric terms
}
\description{
Computes confidence intervals for one or more parameters in a fitted model.
This is a method specific to the \code{"gam"} class from package
\code{"mgcv"}.
}
\examples{
set.seed(101)
library(dplyr)
library(mgcv)
dat <- data.frame(x = runif(n=100),
                  y = runif(n=100)) \%>\%
  dplyr::mutate(z = rnorm(n=100,
                   mean = 1 - 2*x - sin(2*pi*y),
                   sd = 0.1))

fit1 <- gam(data=dat, z ~ y + s(x))

confint(fit1)
confint(fit1, parm="y", level=0.8)

}
