% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_smooths.R
\name{tidy_smooths}
\alias{tidy_smooths}
\title{Tidy Data Frame of Smooth Terms}
\usage{
tidy_smooths(object, dimension = 1, level = 0.95)
}
\arguments{
\item{object}{a fitted model object of class \code{"gam"}.}

\item{dimension}{the dimension of the smooths that are desired for extraction.}

\item{level}{the confidence level required.}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing, all
parameters are considered (not yet implemented).}

\item{...}{not implemented}
}
\value{
A tidy data frame containing parameter names, estimates and
confidence intervals for non-parametric terms
}
\description{
Returns a tidy data frame containing the mean and confidence intervals
for fitted smooths from a model of class \code{"gam"} from package
\code{"mgcv"}.
}
\examples{
set.seed(101)
library(dplyr)
library(mgcv)
library(mgcv.helper)
dat <- data.frame(x = runif(n=100),
                  y = runif(n=100)) \%>\%
  dplyr::mutate(z = rnorm(n=100,
                   mean = 1 - 2*x - sin(2*pi*y),
                   sd = 0.1))

fit1 <- gam(data=dat, z ~ y + s(x))

tidy_smooths(fit1, dimension=1)



}
