% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.mgcv.R
\name{vif.mgcv}
\alias{vif.mgcv}
\title{Variance Inflation Factor}
\usage{
vif.mgcv(mod)
}
\arguments{
\item{mod}{An object of class mgcv}
}
\value{
VIF.df A data frame consisting of the VIF values for each parametric
term in a fitted Generalised Additive Model
}
\description{
This function takes a fitted mgcv model object and returns a data frame of
variance inflation factors
}
